rename sam2 files
