+++ sam3/model/sam3_video_base.py
@@ -21,10 +21,7 @@ from sam3.logger import get_logger
 from sam3.model.data_misc import BatchedDatapoint
 from sam3.model.model_misc import NestedTensor
 from sam3.model.nms_utils import mask_iou
-from sam3.model.video_tracking_with_prompt_utils import (
-    fill_holes_in_mask_scores,
-    mask_to_box,
-)
+from sam3.model.sam3_tracker_utils import fill_holes_in_mask_scores, mask_to_box
 from sam3.train.masks_ops import rle_encode
